{{cookiecutter.app_name}}/requirements/prod.txt
# Everything needed in production

# Flask
Flask==1.0.2
MarkupSafe==1.0
Werkzeug==0.14.1
Jinja2==2.10
itsdangerous==1.1.0
click>=5.0

# Database
flask-mongoengine-0.9.5
mongoengine-0.16.3
pymongo-3.7.2

# Migrations
Flask-Migrate==2.2.1

# Forms
Flask-WTF==0.14.2
WTForms==2.2.1

# Deployment
gunicorn>=19.1.1

# Webpack
flask-webpack==0.1.0

# Auth
Flask-Login==0.4.1
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar==0.10.1

# Environment variable parsing
environs==4.0.0
